requirements.txt
numpy~=1.26.4
flask~=3.0.3
pillow~=10.4.0
tensorflow~=2.17.0